+++ gdetect/dt.cc
@@ -13,6 +13,7 @@
 #include <math.h>
 #include <sys/types.h>
 #include "mex.h"
+#include <cstdint>
 
 /*
  * Generalized distance transforms.
@@ -59,7 +60,7 @@ void mexFunction(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[]) {
   if (mxGetClassID(prhs[0]) != mxDOUBLE_CLASS)
     mexErrMsgTxt("Invalid input");
 
-  const int *dims = mxGetDimensions(prhs[0]);
+  const size_t *dims = mxGetDimensions(prhs[0]);
   double *vals = (double *)mxGetPr(prhs[0]);
   double ax = mxGetScalar(prhs[1]);
   double bx = mxGetScalar(prhs[2]);
